+++ Runtime/Scripts/Components/DontDestroyOnLoadScript.cs
@@ -7,6 +7,8 @@
  * 
  * Additional Comments:
  *		File Line Length: 120
+ * 
+ *      NOTE: DontDestroyOnLoadScript requires the MonoBehaviour's GameObject to have a unique name
  */
 
 using System.Collections;
@@ -18,11 +20,34 @@ namespace Kokowolo.Utilities
     public class DontDestroyOnLoadScript : MonoBehaviour
     {
         /************************************************************/
+        #region Fields
+
+        [Tooltip("whether only one DontDestroyOnLoadScript GameObject can exist with this name")]
+        [SerializeField] private bool isSingletonInstance = true;
+
+        private static List<string> singletonInstanceNames = new List<string>();
+
+        #endregion
+        /************************************************************/
         #region Functions
+
+        private void Awake()
+        {
+            if (isSingletonInstance && HasSingletonInstanceWithName(name)) 
+            {
+                Destroy(gameObject);
+            }
+        }
 
         private void Start()
         {
             DontDestroyOnLoad(this);
+            if (isSingletonInstance) singletonInstanceNames.Add(name);
+        }
+
+        private static bool HasSingletonInstanceWithName(string name)
+        {
+            return singletonInstanceNames.Contains(name);
         }
         
         #endregion