+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.0.30] - 2023-01-25
+### Added
+* setting field to DontDestroyOnLoadScript to determine whether only one DontDestroyOnLoadScript GameObject can exist with its name
+### Changed
+* DontDestroyOnLoadScript to track unique singleton instance DontDestroyOnLoadScript GameObjects
+
 ## [0.0.29] - 2023-01-24
 ### Added
 * a simple DontDestroyOnLoadScript to Components folder